requirements.txt
pytest
pylint
python-dotenv
requests
ipykernel
altair
streamlit
awswrangler
logging
pandas
altair
requests_mock